network: merge server and router
