+++ ciri/node/BUILD
@@ -63,7 +63,6 @@ cc_library(
         ":recent_seen_bytes_cache",
         ":tips_cache",
         "//ciri/node/network:router_shared",
-        "//ciri/node/network:tcp_server_shared",
         "//ciri/node/pipeline:broadcaster_shared",
         "//ciri/node/pipeline:processor_shared",
         "//ciri/node/pipeline:responder_shared",
@@ -81,7 +80,6 @@ cc_library(
         ":node_shared",
         "//ciri:core_shared",
         "//ciri/node/network:router",
-        "//ciri/node/network:tcp_server",
         "//ciri/node/pipeline:broadcaster",
         "//ciri/node/pipeline:processor",
         "//ciri/node/pipeline:responder",
