+++ ciri/node/network/router.h
@@ -9,9 +9,13 @@
 #define __CIRI_NODE_NETWORK_ROUTER_H__
 
 #include "utarray.h"
+#include "uv.h"
 
 #include "ciri/node/network/neighbor.h"
+#include "ciri/node/protocol/type.h"
+#include "common/errors.h"
 #include "utils/handles/rw_lock.h"
+#include "utils/handles/thread.h"
 
 #define NEIGHBORS_FOREACH(neighbors, neighbor)                              \
   for (neighbor = (neighbor_t *)utarray_front(neighbors); neighbor != NULL; \
@@ -21,7 +25,14 @@
 extern "C" {
 #endif
 
+// Forward declarations
+typedef struct node_s node_t;
+
 typedef struct router_s {
+  // Metadata
+  bool running;
+  thread_handle_t thread;
+  // Data
   node_t *node;
   UT_array *neighbors;
   rw_lock_handle_t neighbors_lock;
@@ -36,6 +47,24 @@ typedef struct router_s {
  * @return a status code
  */
 retcode_t router_init(router_t *const router, node_t *const node);
+
+/**
+ * Starts a router
+ *
+ * @param[in,out] router  The router
+ *
+ * @return a status code
+ */
+retcode_t router_start(router_t *const router);
+
+/**
+ * Stops a router
+ *
+ * @param[in,out] router  The router
+ *
+ * @return a status code
+ */
+retcode_t router_stop(router_t *const router);
 
 /**
  * Destroys a router
@@ -108,8 +137,8 @@ neighbor_t *router_neighbor_find_by_endpoint_values(router_t *const router, char
  *
  * @return a pointer to the neighbor if found, NULL otherwise
  */
-retcode_t router_neighbor_read_handshake(router_t *const router, char const *const ip, uint16_t const port,
-                                         void const *const buf, size_t const nread, neighbor_t **const neighbor);
+retcode_t router_read_handshake(router_t *const router, char const *const ip, uint16_t const port,
+                                void const *const buf, size_t const nread, neighbor_t **const neighbor);
 
 /**
  * Reads a buffer from a neighbor
@@ -121,8 +150,21 @@ retcode_t router_neighbor_read_handshake(router_t *const router, char const *con
  *
  * @return a status code
  */
-retcode_t router_neighbor_read(router_t *const router, neighbor_t *const neighbor, void const *const buffer,
-                               size_t const buffer_size);
+retcode_t router_read(router_t *const router, neighbor_t *const neighbor, void const *const buffer,
+                      size_t const buffer_size);
+
+/**
+ * Writes data to a stream
+ *
+ * @param[in] stream      The stream
+ * @param[in] type        The packet type
+ * @param[in] buffer      The buffer
+ * @param[in] buffer_size The buffer size
+ *
+ * @return a status code
+ */
+retcode_t router_write(uv_stream_t *const stream, packet_type_t const type, void *const buffer,
+                       uint16_t const buffer_size);
 
 /**
  * Attempts to reconnect all disconnected neighbors
@@ -131,7 +173,26 @@ retcode_t router_neighbor_read(router_t *const router, neighbor_t *const neighbo
  *
  * @return a status code
  */
-retcode_t router_neighbors_reconnect_attempt(router_t *const router);
+retcode_t router_reconnect_attempt(router_t *const router);
+
+/**
+ * Resolves a domain name into an IP address
+ *
+ * @param[in]   domain  The domain name
+ * @param[out]  ip      The IP address
+ *
+ * @return a status code
+ */
+retcode_t router_resolve(char const *const domain, char *const ip);
+
+/**
+ * Initiates a connection with a neighbor
+ *
+ * @param[in] neighbor The neighbor
+ *
+ * @return a status code
+ */
+retcode_t router_connect(neighbor_t *const neighbor);
 
 #ifdef __cplusplus
 }
